requirements.txt
kafka-python
requests
python-snappy
opensearch-py